+++ helpers/getv.js
@@ -1,0 +1,5 @@
+import pkg from '../package.json' with { type: "json" };
+
+const version = pkg.version ?? pkg.default.version;
+
+export default version;